LayoutTests/fast/eventloop/queue-task-across-frames-in-file-uri-expected.txt

This tests the order by which tasks are scheduled across documents of file URI documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS logs.join(", ") is "1, 2, 3, 4, 5, 6"
PASS internals.hasSameEventLoopAs(frame1.contentWindow) is true
PASS internals.hasSameEventLoopAs(frame2.contentWindow) is true
PASS successfullyParsed is true

TEST COMPLETE
